Split accounting.py into two files
