+++ rotkehlchen/constants.py
@@ -4,3 +4,5 @@ BTC_BCH_FORK_TS = 1501593374  # 2017-08-01 13:16:14 UTC
 SUPPORTED_EXCHANGES = ['kraken', 'poloniex', 'bittrex', 'binance']
 ROTKEHLCHEN_SERVER_TIMEOUT = 5
 ALL_REMOTES_TIMEOUT = 20
+
+YEAR_IN_SECONDS = 31536000  # 60 * 60 * 24 * 365
